+++ CHANGELOG.md
@@ -520,6 +520,7 @@ autorelease pool.
 
 
 
+
 ### Version 0.11.1 beta (2014-03-19)
 
  - Simplified implementation of class method `all` and `any`.