+++ CHANGELOG.md
@@ -527,3 +527,12 @@ autorelease pool.
 
  - Improved setup for iOS Unit Tests
 
+
+### Version 0.11.2 beta (2014-04-01)
+
+ - Fixed a bug in method `runLoopWait`
+
+ - Minimum Deployment target for Mac OS X is now 10.8
+ 
+ 
+